+++ share/templates/share/header.html
@@ -24,6 +24,11 @@
     })(window,document,'//static.hotjar.com/c/hotjar-','.js?sv=');
 </script>
 <style>
+.suggest_form {
+    padding-right: 10%;
+    padding-left: 15%;
+    padding-top:6%;
+}
 .dictionary-search-button {
     background-color: #5DA984;; /* Green */
     border: none;
