+++ word/templates/word/modal_suggest_word.html
@@ -50,7 +50,7 @@
                   {% if lan.proficiency == "advanced" or lan.proficiency == "fluent" or lan.proficiency == "native" %}
 
                     <div>
-                      <label class="col-sm-2" for="description_short_{{ lan.language.name }}">In {% trans lan.language.name %}</label>
+                      <label class="col-sm-2 control-label" for="description_short_{{ lan.language.name }}">In {% trans lan.language.name %}</label>
                       <div class="col-sm-10">
                         <textarea class="form-control" id="description_short_{{ lan.language.name }}"
                                 name="desc_short_{{ lan.language.name }}"></textarea>
@@ -112,8 +112,9 @@
               </div>
 
               <div class="form-group">
-                <a class="col-sm-2 link">advanced</a>
+                <a class="col-sm-2 link control-label">advanced</a>
               </div>
+               <br>
 
               <div class="form-group">
                 <div class="col-xs-offset-1 col-xs-3">
