+++ word/views.py
@@ -1,6 +1,7 @@
 
 
 import re
+import logging
 
 from django.core.files.storage import FileSystemStorage
 from django.urls import reverse_lazy
@@ -13,6 +14,10 @@ from django.contrib.auth.decorators import login_required
 from .models import Word, Description, WordVersion, WordLocation, Tag
 from .forms import WordForm
 from language.models import Language
+from user.models import Profile, UserLanguage
+
+
+logger = logging.getLogger(__name__)
 
 
 class WordView(DetailView):
@@ -42,12 +47,32 @@ class SuggestView(TemplateView):
     template_name = 'word/suggest.html'
     http_method_names = ['get', 'post']
 
+    def get(self, request, *args, **kwargs):
+        context = self.get_context_data(**kwargs)
+
+        try:
+            user_profile = Profile.objects.get(user=request.user)
+        except TypeError:
+            logger.info("user does not have profile")
+        else:
+            user_languages = UserLanguage.objects.filter(user=user_profile)
+            user_moderator = False
+            for user_language in user_languages:
+                if user_language.is_moderator:
+                    user_moderator = True
+                    break
+            context["user_moderator"] = user_moderator
+            context["user_languages"] = user_languages
+
+            default_variant = get_highest_language_proficiency(user_languages)
+            context["default_variant"] = default_variant
+
+        return self.render_to_response(context)
+
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-
-        context.update({
-            'versions': WordVersion.objects.all(),
-        })
+        context["tags"] = Tag.objects.all()
+        context["synonyms"] = Word.objects.all()
         return context
 
     def post(self, request, *args, **kwargs):
@@ -147,6 +172,38 @@ class SuggestView(TemplateView):
         descriptions = list(descriptions.values())
         Description.objects.bulk_create(descriptions)
         return descriptions
+
+
+def get_highest_language_proficiency(user_languages):
+    variants = dict()
+    # variants = {
+    #             "native": ["es_es"],
+    #             "fluent": ["eng_eng", "port_port"],
+    #               }
+
+    for user_language in user_languages:
+        proficiency = user_language.proficiency
+        if proficiency not in variants:
+            variants[proficiency] = list()
+        default_variant = user_language.language.default_variant
+        variants[proficiency].append(default_variant)
+
+    levels = [
+        'native',
+        'fluent',
+        'advanced',
+        'intermediate',
+        'novice',
+        'beginner',
+    ]
+
+    for level in levels:
+        try:
+            default_variant = variants[level][0]
+        except (KeyError, IndexError):
+            # there is no level in variants or variant[level] is empty
+            continue
+        return default_variant
 
 
 class EditView(UpdateView):
